distribution/requirements.txt
pycrypto >= 1.6.1
